backend/requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.1
fastapi>=0.116.0
uvicorn>=0.35.0
python-multipart>=0.0.20
PyPDF2>=3.0.0
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.1
PyPDF2>=3.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

